hw_1/hw1_2/requirements.txt
GitPython==3.1.9